+++ src/constants/index.js
@@ -1,0 +1,12 @@
+const ERROR_MESSAGE = {
+  InputError: '올바르지 못한 입력 값 입니다.',
+};
+
+const CONSOLE_MESSAGE = {
+  StartGame: '숫자 야구 게임을 시작합니다.',
+  GetInputAnswer: '숫자를 입력해주세요 : ',
+  EndGame: '3개의 숫자를 모두 맞히셨습니다! 게임 종료',
+  SelectReplay: '게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.\n',
+};
+
+module.exports = {ERROR_MESSAGE, CONSOLE_MESSAGE};
